TES4: Do not mark SPLO as sorted RRR
This is a workaround - the CS does seem to sort these, but we have no
idea by what it sorts them. The only thing we've figured out so far is
that it always sorts LVSPs after SPELs - which we can't do without
loading the masters (ref # 282 and # 577 <--- RRR). So for now just
don't do any sorting and keep the order created by the CS/CK or xEdit.