Replace our usage of distuils.version RRR
By creating a custom LooseVersion class inspired by distuils' class. But
ours is much better (e.g. unicode support) and has tests to go with it.
The reason to do this is because distutils is getting deprecated in
Python 3.10 and wil be removed at some point in the future.

The only ugly wart here is that we have to copy-paste it into utils.py
so that the build script can use it. Not sure what to do about that.

Under # 619 <--- RRR