Show in-game plugin indices on the Saves tab RRR
We currently show the 'true' in-game indices on the Mods tab (as the
'Indices' column). ESL are given the proper 'FE XYZ' index with this
column, letting you see what index they'll actually have ingame. The
Saves tab should show this information for the masters in a save as
well.

Bit more of a journey than I thought it would be - had to create a way
to 'ban' columns so that the new Indices column wouldn't bleed over into
the masters view on the Mods tab and I obviously had to split save
header masters into two lists - one for regular plugins and one for
ESLs.

Closes # 624 <--- RRR