squash! Unwrap pickle dicts
DataTable must go - profiles are mostly encapsulated in saveInfos, still
it is not clear how they are related to initialization.getLocalSaveDirs.
Should we clean up and maybe avoid the setdefault by making sure we have
all the keys and only those? Anyway renames with the same number of chars
is what we do best.